+++ apps/wizard/app_pages/indicator_upgrade/charts_update.py
@@ -10,9 +10,9 @@ from structlog import get_logger
 
 import etl.grapher_model as gm
 from apps.chart_sync.admin_api import AdminAPI
-from apps.wizard.utils import set_states, st_toast_error
+from apps.wizard.utils import set_states
 from apps.wizard.utils.cached import get_grapher_user_id
-from apps.wizard.utils.components import st_wizard_page_link
+from apps.wizard.utils.components import st_toast_error, st_wizard_page_link
 from apps.wizard.utils.db import WizardDB
 from etl.config import OWID_ENV
 from etl.helpers import get_schema_from_url
