+++ apps/wizard/etl_steps_old/grapher.py
@@ -8,7 +8,7 @@ from typing_extensions import Self
 
 from apps.utils.files import add_to_dag, generate_step_to_channel
 from apps.wizard import utils
-from apps.wizard.etl_steps_old.utils import ADD_DAG_OPTIONS, COOKIE_GRAPHER, MD_GRAPHER
+from apps.wizard.etl_steps_old.utils import ADD_DAG_OPTIONS, COOKIE_GRAPHER, MD_GRAPHER, render_responsive_field_in_form
 from apps.wizard.utils.components import config_style_html, preview_file, st_wizard_page_link
 from etl.paths import DAG_DIR
 
@@ -110,7 +110,6 @@ with st.container(border=True):
 
 # SIDEBAR
 with st.sidebar:
-    # utils.warning_metadata_unstable()
     # CONNECT AND
     if APP_STATE.args.run_checks:
         with st.expander("**Environment checks**", expanded=True):
@@ -186,7 +185,7 @@ with form_widget.form("grapher"):
 
 
 # Render responsive namespace field
-utils.render_responsive_field_in_form(
+render_responsive_field_in_form(
     key="namespace",
     display_name="Namespace",
     field_1=namespace_field[0],
